add ability to remove/clear languages
added ability to remove and clear languages. removing a language removes it from the vector entirely; all indices above the removed language decrease by one.
clearing a language empties the language - removes all keys and strings - but the language remains available but is empty (in the same state as after adding the language)